refactor: move checks/ into mtui/update_workflow/
Step 5c of the package reorganization (see
PROPOSED_CODE_FILE_REORGANIZATION_PLAN.md). Moves the checks package
under its new home:

  mtui/checks/ -> mtui/update_workflow/checks/

Importer rewrites:
  mtui/target/target.py: from ..checks import ...
    -> from ..update_workflow.checks import ...
  tests/test_checks_{downgrade,install,prepare,update}.py:
    from mtui.checks.X import ... -> from mtui.update_workflow.checks.X import ...

Within the moved package, two-dot relative imports to ..support.* and
..cli.* were rewritten to three-dot ...support.* / ...cli.* so they
still reach those packages after the extra level.

Logger names (getLogger("mtui.checks.X")) intentionally left at their
original strings; logger names are stable identifiers independent of
the import path, matching the precedent from the support/ move.

Doc comment in target.py updated to point at the new module paths.

All four CI gates green; no behaviour change.
