+++ mtui/update_workflow/checks/install.py
@@ -3,7 +3,7 @@
 from collections.abc import Callable
 from logging import getLogger
 
-from ..support.exceptions import UpdateError
+from ...support.exceptions import UpdateError
 
 logger = getLogger("mtui.checks.install")
