+++ mtui/update_workflow/checks/prepare.py
@@ -1,11 +1,11 @@
-"""Defines checks to be performed after a downgrade action."""
+"""Defines checks to be performed after a prepare action."""
 
 from collections.abc import Callable
 from logging import getLogger
 
-from ..support.exceptions import UpdateError
+from ...support.exceptions import UpdateError
 
-logger = getLogger("mtui.checks.downgrade")
+logger = getLogger("mtui.checks.prepare")
 
 
 def zypper(hostname: str, stdout: str, stdin: str, stderr: str, exitcode: int) -> None:
@@ -25,16 +25,19 @@ def zypper(hostname: str, stdout: str, stdin: str, stderr: str, exitcode: int) -
     if "A ZYpp transaction is already in progress." in stderr:
         logger.critical(
             '%s: command "%s" failed:\nstdin:\n%s\nstderr:\n%s',
-        )
-        raise UpdateError(hostname, "update stack locked")
-    if "System management is locked" in stderr:
-        logger.critical(
-            '%s: command "%s" failed:\nstdout:\n%s\nstderr:\n%s',
             hostname,
             stdin,
             stdout,
             stderr,
-            exitcode,
+        )
+        raise UpdateError(hostname, "update stack locked")
+    if "System management is locked" in stderr:
+        logger.critical(
+            '%s: command "%s" failed:\nstdin:\n%s\nstderr:\n%s',
+            hostname,
+            stdin,
+            stdout,
+            stderr,
         )
         raise UpdateError("update stack locked", hostname)
     if "(c): c" in stdout:
@@ -44,15 +47,19 @@ def zypper(hostname: str, stdout: str, stdin: str, stderr: str, exitcode: int) -
             stdout,
         )
         raise UpdateError("Dependency Error", hostname)
-    if exitcode == 104:
-        logger.critical("%s: zypper returned with errorcode 104:\n%s", hostname, stderr)
-        raise UpdateError("Unspecified Error", hostname)
-    if exitcode == 106:
-        logger.warning("%s: zypper returned with errocode 106:\n%s", hostname, stderr)
+    if "Error:" in stderr:
+        logger.critical(
+            '%s: command "%s" failed:\nstdin:\n%s\nstderr:\n%s',
+            hostname,
+            stdin,
+            stdout,
+            stderr,
+        )
+        raise UpdateError("RPM Error", hostname)
 
 
-#: A dictionary that maps system configurations to downgrade check functions.
-downgrade_checks: dict[tuple[str, bool], Callable[[str, str, str, str, int], None]] = {
+#: A dictionary that maps system configurations to prepare check functions.
+prepare_checks: dict[tuple[str, bool], Callable[[str, str, str, str, int], None]] = {
     ("11", False): zypper,
     ("12", False): zypper,
     ("15", False): zypper,
